+++ src/main.tsx
@@ -20,6 +20,7 @@ import { Project, LoadError } from "./typings";
 import "./styles/_style.scss";
 import { getProjectWindowLabel } from "./utils/getProjectWindowLabel";
 import { createWebDir } from "./fs/createWebDir";
+import { setAccentColor } from "./utils/setAccentColor";
 import { listen } from "@tauri-apps/api/event";
 import loadProject from "./fs/loadProject";
 import { accentColor } from "tauri-plugin-accent-color";
@@ -29,18 +30,7 @@ const appWindow = getCurrentWebviewWindow();
 async function init() {
   // Set accent color
   accentColor.subscribe((color) => {
-    let accentColorEl = document.querySelector("#accent-color-styles");
-    if (!accentColorEl) {
-      accentColorEl = document.createElement("style");
-      accentColorEl.setAttribute("id", "accent-color-styles");
-    }
-    // TODO: Fill out accent color values
-    accentColorEl.textContent = `
-	  :root { 
-		  --accent-color: ${color};
-	  }
-	  `;
-    document.head.appendChild(accentColorEl);
+    setAccentColor(color);
   });
   // Window focus
   listen("tauri://focus", async () => {